refactor: transport is not a converter
Make sure we differentiate between a converter and a transport.
