+++ src/libecalc/energy/energy_units/__init__.py
@@ -1,7 +1,8 @@
 from .consumers import BaseLoad, Compressor, DieselConsumer, Flare, Pump, SampledFuelConsumer, SampledPowerConsumer
-from .converters import ElectricalCable, ElectricalMotor, GasTurbine, GeneratorSet, Transporter
+from .converters import ElectricalMotor, GasTurbine, GeneratorSet
 from .junction import ElectricalBus, FuelGasManifold, Junction
 from .sources import DieselSupply, FuelGasSource, OffshoreWind, OnshoreGrid
+from .transporter import ElectricalCable, Transporter
 
 __all__ = [
     "BaseLoad",
