+++ src/libecalc/energy/energy_units/converters.py
@@ -1,63 +1,9 @@
-import abc
 from collections.abc import Callable
 from typing import Final
 
-from libecalc.energy.energy_types import ElectricalPower, Energy, FuelGasRate, MechanicalPower
+from libecalc.energy.energy_types import ElectricalPower, FuelGasRate, MechanicalPower
 from libecalc.energy.energy_unit import EnergyUnitId
 from libecalc.energy.provider import Converter
-
-
-class Transporter(Converter):
-    """Moves energy without changing form, possibly with loss.
-
-    Subclasses set energy_type — input and output are always the same.
-    """
-
-    @classmethod
-    @abc.abstractmethod
-    def get_energy_type(cls) -> type[Energy]: ...
-
-    @classmethod
-    def get_input_energy_type(cls) -> type[Energy]:
-        return cls.get_energy_type()
-
-    @classmethod
-    def get_output_energy_type(cls) -> type[Energy]:
-        return cls.get_energy_type()
-
-
-class ElectricalCable(Transporter):
-    """Electrical cable with transmission loss (e.g. subsea cable from shore)."""
-
-    def __init__(
-        self, name: str, max_power: float, loss_fraction: float = 0.0, energy_unit_id: EnergyUnitId | None = None
-    ) -> None:
-        self._name = name
-        self._max_power = max_power
-        self._loss_fraction = loss_fraction
-        self._id: Final[EnergyUnitId] = energy_unit_id or ElectricalCable._create_id()
-
-    @classmethod
-    def get_energy_type(cls) -> type[ElectricalPower]:
-        return ElectricalPower
-
-    def get_id(self) -> EnergyUnitId:
-        return self._id
-
-    def get_name(self) -> str:
-        return self._name
-
-    def get_max_power(self) -> float:
-        return self._max_power
-
-    def get_loss_fraction(self) -> float:
-        return self._loss_fraction
-
-    def capacity(self) -> ElectricalPower | None:
-        return ElectricalPower(self._max_power)
-
-    def get_input_energy(self, output_energy: ElectricalPower) -> ElectricalPower:
-        return ElectricalPower(output_energy.value / (1 - self._loss_fraction))
 
 
 class GeneratorSet(Converter):
